--- example/__init__.py
@@ -1,19 +1,0 @@
-"""This is the "example" module.
-
-The example module supplies one function: compute(). For example,
-
->>> compute(3)
-3
-"""
-
-
-def compute(x):
-    """Functon compute returns evaluation of expression using argument x.
-
-    Args:
-        - x - Input of the function
-
-    Returns:
-        - output - Output of the function
-    """
-    return x * x - 2 * x
